requirements.txt
openpyxl==3.1.5
python-docx==1.2.0
python-pptx==1.0.2
matplotlib==3.10.8
boto3==1.42.89
